samples/cdk-lambda-strands-durable/lambda/demo-one/requirements.txt
strands-agents==1.18.0
strands-agents-tools==0.2.16
aws-durable-execution-sdk-python==1.0.0.post1